1. Created new namespace for predictions 2. Reverted code to keep token_authenticator methods as static
